Prettier u7

+++ src/util/util/Config.ts
@@ -42,9 +42,7 @@ export class Config {
 		} else {
 			console.log(`[Config] Using CONFIG_PATH rather than database`);
 			if (existsSync(process.env.CONFIG_PATH)) {
-				const file = JSON.parse(
-					(await fs.readFile(process.env.CONFIG_PATH)).toString(),
-				);
+				const file = JSON.parse((await fs.readFile(process.env.CONFIG_PATH)).toString());
 				config = file;
 			} else config = new ConfigValue();
 			pairs = generatePairs(config);
@@ -58,7 +56,7 @@ export class Config {
 		await this.set(config);
 		validateFinalConfig(config);
 		return config;
-	};
+	}
 	public static get() {
 		if (!config) {
 			// If we haven't initialised the config yet, return default config.
@@ -70,13 +68,13 @@ export class Config {
 		}
 
 		return config;
-	};
+	}
 	public static set(val: Partial<ConfigValue>) {
 		if (!config || !val) return;
 		config = OrmUtils.mergeDeep(config);
 
 		return applyConfig(config);
-	};
+	}
 }
 
 // TODO: better types
@@ -98,13 +96,12 @@ const generatePairs = (obj: object | null, key = ""): ConfigEntity[] => {
 
 async function applyConfig(val: ConfigValue) {
 	if (process.env.CONFIG_PATH)
-		if (!process.env.CONFIG_READONLY)
-			await fs.writeFile(overridePath, JSON.stringify(val, null, 4));
+		if (!process.env.CONFIG_READONLY) await fs.writeFile(overridePath, JSON.stringify(val, null, 4));
 		else console.log("[WARNING] JSON config file in use, and writing is disabled! Programmatic config changes will not be persisted, and your config will not get updated!");
 	else {
 		const pairs = generatePairs(val);
 		// keys are sorted to try to influence database order...
-		await Promise.all(pairs.sort((x, y) => x.key > y.key ? 1 : -1).map((pair) => pair.save()));
+		await Promise.all(pairs.sort((x, y) => (x.key > y.key ? 1 : -1)).map((pair) => pair.save()));
 	}
 	return val;
 }
@@ -122,8 +119,7 @@ function pairsToConfig(pairs: ConfigEntity[]) {
 		let i = 0;
 
 		for (const key of keys) {
-			if (!isNaN(Number(key)) && !prevObj[prev]?.length)
-				prevObj[prev] = obj = [];
+			if (!isNaN(Number(key)) && !prevObj[prev]?.length) prevObj[prev] = obj = [];
 			if (i++ === keys.length - 1) obj[key] = p.value;
 			else if (!obj[key]) obj[key] = {};
 
@@ -143,7 +139,7 @@ const validateConfig = async () => {
 
 	for (const row in config) {
 		// extension methods...
-		if(typeof config[row] === "function") continue;
+		if (typeof config[row] === "function") continue;
 
 		try {
 			const found = await ConfigEntity.findOne({
@@ -152,11 +148,7 @@ const validateConfig = async () => {
 			if (!found) continue;
 			config[row] = found;
 		} catch (e) {
-			console.error(
-				`Config key '${config[row].key}' has invalid JSON value : ${
-					(e as Error)?.message
-				}`,
-			);
+			console.error(`Config key '${config[row].key}' has invalid JSON value : ${(e as Error)?.message}`);
 			hasErrored = true;
 		}
 	}
@@ -164,9 +156,7 @@ const validateConfig = async () => {
 	console.log("[Config] Total config load time:", new Date().getTime() - totalStartTime.getTime(), "ms");
 
 	if (hasErrored) {
-		console.error(
-			"[Config] Your config has invalid values. Fix them first https://docs.spacebar.chat/setup/server/configuration",
-		);
+		console.error("[Config] Your config has invalid values. Fix them first https://docs.spacebar.chat/setup/server/configuration");
 		process.exit(1);
 	}
 
@@ -194,15 +184,13 @@ function validateFinalConfig(config: ConfigValue) {
 		}
 	}
 
-	assertConfig("api_endpointPublic", v => v != null, "A valid public API endpoint URL, ex. \"http://localhost:3001/api/v9\"");
-	assertConfig("cdn_endpointPublic", v => v != null, "A valid public CDN endpoint URL, ex. \"http://localhost:3003/\"");
-	assertConfig("cdn_endpointPrivate", v => v != null, "A valid private CDN endpoint URL, ex. \"http://localhost:3003/\" - must be routable from the API server!");
-	assertConfig("gateway_endpointPublic", v => v != null, "A valid public gateway endpoint URL, ex. \"ws://localhost:3002/\"");
+	assertConfig("api_endpointPublic", (v) => v != null, 'A valid public API endpoint URL, ex. "http://localhost:3001/api/v9"');
+	assertConfig("cdn_endpointPublic", (v) => v != null, 'A valid public CDN endpoint URL, ex. "http://localhost:3003/"');
+	assertConfig("cdn_endpointPrivate", (v) => v != null, 'A valid private CDN endpoint URL, ex. "http://localhost:3003/" - must be routable from the API server!');
+	assertConfig("gateway_endpointPublic", (v) => v != null, 'A valid public gateway endpoint URL, ex. "ws://localhost:3002/"');
 
 	if (hasErrors) {
-		console.error(
-			"[Config] Your config has invalid values. Fix them first https://docs.spacebar.chat/setup/server/configuration",
-		);
+		console.error("[Config] Your config has invalid values. Fix them first https://docs.spacebar.chat/setup/server/configuration");
 		console.error("[Config] Hint: if you're just testing with bundle (`npm run start`), you can set all endpoint URLs to [proto]://localhost:3001");
 		process.exit(1);
 	} else console.log("[Config] Configuration validated successfully.");
